[REF] web,html_editor: call useColorPicker to open ColorPicker
The goal of this PR is to create a hook that makes it easy to display
a ColorPicker. Now, to open a ColorPicker, you need to call useColorPicker
with the ref of the target and a function to get props when you open it.

Since this PR, the ColorPicker refers to the popover with the 3 tabs (solid, custom and gradient).
The old ColorPicker component is renamed CustomColorPicker.
