+++ addons/web/static/src/core/color_picker/gradient_picker/gradient_picker.xml
@@ -1,5 +1,5 @@
 <templates xml:space="preserve">
-    <t t-name="html_editor.GradientPicker">
+    <t t-name="web.GradientPicker">
         <div class="d-flex flex-column">
 
             <div class="align-items-center d-flex justify-content-between my-2 o_type_row">
