+++ addons/web_editor/static/src/js/wysiwyg/widgets/color_palette.js
@@ -1,4 +1,4 @@
-import { ColorPicker } from "@web/core/color_picker/color_picker";
+import { CustomColorPicker as ColorPicker } from "@web/core/color_picker/custom_color_picker/custom_color_picker";
 import customColors from "@web_editor/js/editor/custom_colors";
 import weUtils from "@web_editor/js/common/utils";
 import {
